% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.R
\name{funnel}
\alias{funnel}
\title{A funnel plot}
\usage{

  funnel(json = "~/Documents/SoftwareDevelopment/VikingCodeSchool/final_project/openMeta/src/newStudiesJSON.json")
}
\arguments{
\item{df}{json array}
}
\description{
This function allows you to generate funnel plots with OpenMeta.
}
\examples{
funnel()
NOTE: for ggplot() to work, "import(ggplot2)" must be in the NAMESPACE file
------------------------------------------------------------------------
DATA FRAME transformations
-----------------------------------------------------------------------
Ensures that only studies selected for analysis are available
------------------------------------------------------------------------
DISPLAY PROPERTIES
-----------------------------------------------------------------------
------------------------------------------------------------------------
PACKAGE FUNCTIONS
-----------------------------------------------------------------------
If your module has any dependencies, it is recommended you require the
functions from those dependencies that you need, below, like so:
functionNeeded <- packageUsed::functionNeeded
ggplot2
metafor
plotly
------------------------------------------------------------------------
CUSTOM MODULE CODE
------------------------------------------------------------------------
Put your primary code below
}
\keyword{ocpu}
