% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedEffects.R, R/simplePlot.R
\name{simplePlot}
\alias{simplePlot}
\title{A funnel plot}
\usage{

  simplePlot(json = "~/Documents/SoftwareDevelopment/VikingCodeSchool/final_project/openMeta/src/dbstudiesJSON.json")


  simplePlot(json = "~/Documents/SoftwareDevelopment/VikingCodeSchool/final_project/openMeta/src/dbstudiesJSON.json")
}
\arguments{
\item{df}{json array}

\item{df}{json array}
}
\description{
This function allows you to generate funnel plots with OpenMeta.

This function was created to allow you to test the OpenCPU service
}
\examples{
simplePlot()
NOTE: for ggplot() to work, "import(ggplot2)" must be in the NAMESPACE file
------------------------------------------------------------------------
DATA FRAME transformations
-----------------------------------------------------------------------
Ensures that only studies selected for analysis are available
------------------------------------------------------------------------
DISPLAY PROPERTIES
-----------------------------------------------------------------------
------------------------------------------------------------------------
PACKAGE FUNCTIONS
-----------------------------------------------------------------------
If your module has any dependencies, it is recommended you require the
functions from those dependencies that you need, below, like so:
functionNeeded <- packageUsed::functionNeeded
------------------------------------------------------------------------
CUSTOM MODULE CODE
------------------------------------------------------------------------
Put your primary code below
simplePlot()
NOTE: for ggplot() to work, "import(ggplot2)" must be in the NAMESPACE file
------------------------------------------------------------------------
DATA FRAME transformations
-----------------------------------------------------------------------
Ensures that only studies selected for analysis are available
------------------------------------------------------------------------
DISPLAY PROPERTIES
-----------------------------------------------------------------------
------------------------------------------------------------------------
PACKAGE FUNCTIONS
-----------------------------------------------------------------------
If your module has any dependencies, it is recommended you require the
functions from those dependencies that you need, below, like so:
functionNeeded <- packageUsed::functionNeeded
------------------------------------------------------------------------
CUSTOM MODULE CODE
------------------------------------------------------------------------
Put your primary code below
}
\keyword{ocpu}
