% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{A test function}
\usage{
test_function(num)
}
\arguments{
\item{num}{Can you type a number?}
}
\description{
This function was created to allow you to test the OpenCPU service
}
\examples{
test_function()
}
\keyword{ocpu}
