% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plot.R
\name{test_plot}
\alias{test_plot}
\title{A test function}
\usage{
test_plot(x, y)
}
\arguments{
\item{x}{x-variable}

\item{y}{y-variable}
}
\description{
This function was created to allow you to test the OpenCPU service
}
\examples{
test_plot()
}
\keyword{ocpu}
