% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlot.R
\name{simplePlot}
\alias{simplePlot}
\title{A test function}
\usage{

  simplePlot(json = "~/Documents/SoftwareDevelopment/VikingCodeSchool/final_project/openMeta/src/dbstudiesJSON.json")
}
\arguments{
\item{df}{json array}
}
\description{
This function was created to allow you to test the OpenCPU service
}
\examples{
simplePlot()
}
\keyword{ocpu}
